requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
matplotlib>=3.7.0
Pillow>=9.4.0
tqdm>=4.65.0
scipy>=1.10.0
PyYAML>=6.0
opencv-python>=4.8.0
gradio>=4.0.0
tensorboard>=2.12.0